fix rightmost columns
It seems that I made a mistake while creating the PCB schematic for the
buenzliboard based on the keyberon-f4 code. I didn't realize there are 13 gpio
pins for the columns (which I still don't understand why that's needed.)

Removing one of the gpio pins from the code brings it in line with how my PCB
is designed.